% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_results.R
\name{collapse_for_poseidon}
\alias{collapse_for_poseidon}
\title{Collapse value for poseidon janno}
\usage{
collapse_for_poseidon(x, field)
}
\arguments{
\item{x}{character. A collapsed, poseidon-formatted string of Library builds or UDG treatments.}

\item{field}{character. The field to format for. One of 'UDG', 'Library_Built'}
}
\value{
character. Either the unique repeated element in the collapsed string, or 'mixed'.
}
\description{
Takes a comples ';' separated value of UDG or Library_Built and returns the appropriate j
anno value to summarise that dataset in the provided field.
}
\examples{
collapse_for_poseidon("ds;ds;ds", "Library_Built") ## Unique repeated library build
collapse_for_poseidon("ds;ss;ds", "Library_Built") ## Non-unique library builds

collapse_for_poseidon("minus;minus;minus", "UDG") ## Unique repeated UDG treatment
collapse_for_poseidon("minus;half;plus", "UDG") ## Non-unique UDG treatment
}
