% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_tables.R
\name{compile_across_lib_results}
\alias{compile_across_lib_results}
\title{Compile library level results to sample level}
\usage{
compile_across_lib_results(x, snp_cutoff = 100)
}
\arguments{
\item{x}{A Tibble as given by \link[eager2poseidon]{compile_eager_result_tables}.}

\item{snp_cutoff}{integer. The minimum number of SNPs, below which nuclear contamination results should be ignored.}
}
\value{
A tibble with the following columns replaced by weighted sums on the sample level: c(Contamination, Contamination_Error, Damage).
}
\description{
Function to calculate sample_level damage and contamination using a weighted mean across libraries
}
