% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_tables.R
\name{compile_eager_result_tables}
\alias{compile_eager_result_tables}
\title{Compile eager results tables}
\usage{
compile_eager_result_tables(
  tsv_table = NULL,
  sexdet_table = NULL,
  snpcov_table = NULL,
  dmg_table = NULL,
  endogenous_table = NULL,
  nuccont_table = NULL,
  contamination_method = "1",
  contamination_algorithm = "ml",
  XX_cutoffs = c(0.7, 1.2, 0, 0.1),
  XY_cutoffs = c(0.2, 0.6, 0.3, 0.6),
  capture_type = NA_character_
)
}
\arguments{
\item{tsv_table}{A tibble with the eager TSV input information as returned by \link[eagerR]{infer_merged_bam_names}.}

\item{sexdet_table}{A tibble with the eager sex determination information as returned by \link[eagerR]{read_sexdet_json}.}

\item{snpcov_table}{A tibble with the eager eigenstrat snp coverage information as returned by \link[eagerR]{read_snp_coverage_json}.}

\item{dmg_table}{A tibble with the eager damageprofiler results, as returned by \link[eagerR]{read_damageprofiler_jsons_from_dir}.}

\item{endogenous_table}{A tibble with the eager endogenous DNA information as returned by \link[eagerR]{read_endorspy_jsons_from_dir}.}

\item{nuccont_table}{A tibble withthe eager nuclear contamination results as returned by \link[eagerR]{read_angsd_cont_json}.}

\item{contamination_method}{The contamination method to keep. Can be either "1" or "2".}

\item{contamination_algorithm}{The estimation algorithm to keep. Can be "ml" or "mom".}

\item{XX_cutoffs}{numeric. A vector containing the minimum x_rate, maximum x_rate, minimum y_rate and maximum y_rate for XX calls.}

\item{XY_cutoffs}{numeric. A vector containing the minimum x_rate, maximum x_rate, minimum y_rate and maximum y_rate for XY calls.}

\item{capture_type}{character. The Capture_Type. Assumes all libraries have the same one. Defaults to NA, which will disable inference of the Capture_Type and return a dummy column.}
}
\value{
A tibble with the compiled information from the provided eager output tables.
}
\description{
Compile together various tables of eager results read into poseidon-compatible tables.
}
