% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_tables.R
\name{create_contamination_note}
\alias{create_contamination_note}
\title{Create contamination note from eager results for poseidon}
\usage{
create_contamination_note(lib_ids, nr_snps, snp_cutoff = 100)
}
\arguments{
\item{lib_ids}{character. The names of the libraries the contamination estimates correspond to.}

\item{nr_snps}{character. The number of SNPs for the contamination estimate.}

\item{snp_cutoff}{integer. The minimum number of SNPs, below which nuclear contamination results should be ignored.}
}
\value{
character. An appropriate Contamination_Note
}
\description{
Create contamination note for the sample, ignoring duplicated library entries.
}
