% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_operations.R
\name{fill_genetic_sex}
\alias{fill_genetic_sex}
\title{Add genetic sex information to the input janno, throwing a warning if any changes happen, so users know to update their .fam/.ind files}
\usage{
fill_genetic_sex(input_janno_table, genetic_sex_table)
}
\arguments{
\item{input_janno_table}{tibble. A standardised janno tibble, as returned by \link[eager2poseidon:standardise_janno]{standardise_janno}.}

\item{genetic_sex_table}{tibble. A tibble containing the Poseidon_Id and Sex of inferred genetic sex of each sample}
}
\value{
tibble. A tibble with filled in Genetic_Sex information.
A warning is printed if any entries have been updated, requesting manual changes to be made to the .fam/.ind files of the poseidon package.
}
\description{
Add genetic sex information to the input janno, throwing a warning if any changes happen, so users know to update their .fam/.ind files
}
