% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_operations.R
\name{fill_in_janno}
\alias{fill_in_janno}
\title{Fill in a janno table}
\usage{
fill_in_janno(input_janno_table, external_results_table, genotype_ploidy)
}
\arguments{
\item{input_janno_table}{A standardised janno tibble, as returned by \link[eager2poseidon:standardise_janno]{standardise_janno}.}

\item{external_results_table}{A tibble with the collected results form pandora and eager, as returned by \link[eager2poseidon:collate_external_results]{collate_external_results}.}

\item{genotype_ploidy}{character. The genotype ploidy of the genotypes in the dataset. Can be either 'diploid' or 'haploid'.}
}
\value{
A tibble with missing values filled in for the fields
Collection_ID, Country, Site, Location, Longitude,
Latitude, Date_C14_Labnr, Date_BC_AD_Start, Date_BC_AD_Stop, Date_BC_AD_Median,
Date_C14_Uncal_BP, Date_C14_Uncal_BP_Err, Date_Type, Nr_Libraries, Capture_Type,
UDG, Library_Built, Damage, Nr_SNPs, Endogenous, Contamination, Contamination_Err
}
\description{
Fill in missing information in a janno table with relevant data from pandora,
an eager TSV and the MultiQC general stats table of an eager run using that TSV.
Any cells that are already filled in the janno table will be kept as is.
}
