% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandora_from_janno.R
\name{get_individual_pandora_data}
\alias{get_individual_pandora_data}
\title{Pull pandora information for specific individual IDs}
\usage{
get_individual_pandora_data(sample_ids, credentials)
}
\arguments{
\item{sample_ids}{character. A vector of individual IDs to pull pandora data for.}

\item{credentials}{character. Path to a credentials file containing four lines listing the database host,
the port of the database server, user and password, respectively.
Passed to \link[sidora.core:get_pandora_connection]{get_pandora_connection}}
}
\value{
A tibble containing the pandora information for individuals present in the janno file, from Site to Sequencing.
}
\description{
Query pandora for information across tabs Site and Sequencing, for a provided set of individual IDs.
}
