% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandora_from_janno.R
\name{import_pandora_data}
\alias{import_pandora_data}
\title{Collect poseidon janno data from pandora}
\usage{
import_pandora_data(sample_ids, credentials, trust_uncalibrated_dates)
}
\arguments{
\item{sample_ids}{character. A vector of the Poseidon_IDs to pull from pandora.}

\item{credentials}{character. Path to a credentials file containing four lines listing the database host,
the port of the database server, user and password, respectively.
Passed to \link[sidora.core:get_pandora_connection]{get_pandora_connection}}

\item{trust_uncalibrated_dates}{logical. Should any uncalibrated dates in pandora be trusted?
If set to TRUE, then \link[janno:quickcalibrate]{quickcalibrate()} is used to calibrate these dates on the fly.}
}
\value{
A tibble containing the poseidon janno fields of
}
\description{
Collect poseidon janno data from pandora
}
