% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_columns.R
\name{infer_genetic_sex}
\alias{infer_genetic_sex}
\title{Infer Poseidon encoded genetic sex from relative X/Y coverage}
\usage{
infer_genetic_sex(
  x_rate,
  y_rate,
  XX_cutoffs = c(0.7, 1.2, 0, 0.1),
  XY_cutoffs = c(0.2, 0.6, 0.3, 0.6)
)
}
\arguments{
\item{x_rate}{double. The relative coverage on the X chromosome.}

\item{y_rate}{double. The relative coverage on the Y chromosome.}

\item{XX_cutoffs}{numeric. A vector containing the minimum x_rate, maximum x_rate, minimum y_rate and maximum y_rate for XX calls.}

\item{XY_cutoffs}{numeric. A vector containing the minimum x_rate, maximum x_rate, minimum y_rate and maximum y_rate for XY calls.}
}
\value{
character. The inferred poseidon-encoded genetic sex.
}
\description{
Infer Poseidon encoded genetic sex from relative X/Y coverage
}
