% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_tables.R
\name{library_unique_weighted_mean}
\alias{library_unique_weighted_mean}
\title{Calculate the weighted mean at the library level}
\usage{
library_unique_weighted_mean(lib_ids, measure, weight, na.rm = T)
}
\arguments{
\item{lib_ids}{character. The names of the libraries the measures correspond to.}

\item{measure}{character. The measures to get a weighted mean for}

\item{weight}{character. The weight of each measure.}

\item{na.rm}{logical. Should NAs be removed from the weighted mean. passed on to stats::weighted.mean(na.rm)}
}
\value{
character. The weighted mean of the unique library-level measures.
}
\description{
This function calculates the weighted mean of the measures provided, given the weights provided.
Should be provided the measures for one Sample_Name/Poseidon_ID at a time.
Unlike stats::weighted.mean, this function first filters for distinct rows, to avoid counting duplicated results
as independent observations.
}
