% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandora_from_janno.R
\name{pandora_from_janno}
\alias{pandora_from_janno}
\title{pandora_from_janno}
\usage{
pandora_from_janno(janno, credentials)
}
\arguments{
\item{janno}{character. Path to the source .janno file. Passed to \link[poseidonR:janno]{read_janno}}

\item{credentials}{character. Path to a credentials file containing four lines listing the database host, the port of the database server, user and password, respectively. Passed to \link[sidora.core:get_pandora_connection]{get_pandora_connection}}
}
\value{
A tibble containing the pandora information for individuals present in the janno file, from Site to Sequencing.
}
\description{
Query pandora for information for all individuals present in a janno file.
}
