% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_results.R
\name{read_eager_tsv}
\alias{read_eager_tsv}
\title{Read content of an eager TSV and convert information to poseidon janno columns}
\usage{
read_eager_tsv(eager_tsv_fn, keep_only = "none")
}
\arguments{
\item{eager_tsv_fn}{character. Path to the TSV file used as input for the eager run.}

\item{keep_only}{character. Can be set to "none", "single", or "double".
If set to 'single; or 'double', will keep only information for libraries with the specified strandedness.
If 'none', all information is retained.}
}
\value{
A tibble with columns for information for janno columns Sample_Name, Nr_Libraries, Capture_Type, UDG,
and Library_Built, and an additional column with the IDs of the libraries included for each sample. The latter
is used to pull only relevant information from the eager results tables.
}
\description{
Read content of an eager TSV and convert information to poseidon janno columns
}
