% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandora_from_janno.R
\name{select_calibration_curve}
\alias{select_calibration_curve}
\title{Infer the correct calibration curve for quickcalibrate, based on the Pandora Latitude}
\usage{
select_calibration_curve(Latitude)
}
\arguments{
\item{Latitude}{numeric. Latitude in decimal degrees.}
}
\value{
A string with the preferred calibration curve (intcal20 for northern hemisphere and shcal20 for the southern hemisphere).
}
\description{
Pick a calibration curve for
}
\examples{
select_calibration_curve(-50) ## Northern hemisphere
select_calibration_curve(50) ## Southern Hemisphere
select_calibration_curve(0) ## 0 defaults to Northern hemisphere
\dontrun{
select_calibration_curve("banana")
}
}
