% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_operations.R
\name{collate_external_results}
\alias{collate_external_results}
\title{Collect Pandora and Eager results}
\usage{
collate_external_results(
  sample_ids,
  eager_tsv_fn,
  general_stats_fn,
  credentials,
  keep_only = "none",
  trust_uncalibrated_dates = F,
  snp_cutoff
)
}
\arguments{
\item{sample_ids}{character. A vector of the Poseidon_IDs to pull from pandora.}

\item{eager_tsv_fn}{character. Path to the TSV file used as input for the eager run.}

\item{general_stats_fn}{character. Path to the MultiQC general stats table.
Can be found in multiqc/multiqc_data/multiqc_general_stats.txt within the specified eager output directory (\code{--outdir}).}

\item{credentials}{character. Path to a credentials file containing four lines listing the database host,
the port of the database server, user and password, respectively.
Passed to \link[sidora.core:get_pandora_connection]{get_pandora_connection}}

\item{keep_only}{character. Can be set to "none", "single", or "double".
If set to 'single; or 'double', will keep only information for libraries with the specified strandedness.
If 'none', all information is retained.}

\item{trust_uncalibrated_dates}{logical. Should any uncalibrated dates in pandora be trusted?
If set to TRUE, then \link[poseidonR:quickcalibrate]{quickcalibrate()} is used to calibrate these dates on the fly.}

\item{snp_cutoff}{integer. The minimum number of SNPs used for nuclear contamination results, for the results to be
considered trustworthy. Any values from libraries with fewer than this number of SNPs are ignored.}
}
\value{
A tibble with the collected results from Pandora and eager.
}
\description{
Collect Pandora and Eager results
}
