% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eager_results.R
\name{read_eager_stats_table}
\alias{read_eager_stats_table}
\title{Extract poseidon information from the general stats table of a run}
\usage{
read_eager_stats_table(general_stats_fn, tsv_data)
}
\arguments{
\item{general_stats_fn}{character. Path to the MultiQC general stats table.
Can be found in multiqc/multiqc_data/multiqc_general_stats.txt within the specified eager output directory (\code{--outdir}).}

\item{tsv_data}{A tibble containing information from the eager TSV. Created with \link[eager2poseidon:read_eager_tsv]{read_eager_tsv}}
}
\value{
A tibble containing the poseidon janno fields of Damage, Nr_SNPs, Endogenous, Contamination,
Contamination_Err, Contamination_Meas, and Contamination_Note
}
\description{
Extract poseidon information from the general stats table of a run
}
